Refactored into separate files for dataclasses and clients
