+++ examples/find_speaker.py
@@ -2,14 +2,14 @@
 import sys, os
 sys.path.append(os.path.abspath(os.path.join(os.path.dirname(__file__), '..')))
 
-import notubiz
-import notubiz.api.speakers
+from notubiz import ApiClient, Configuration
+from notubiz.api.clients import SpeakersClient
 
-configuration = notubiz.Configuration(organisation_id = 686) # Gemeente Eindhoven
+configuration = Configuration(organisation_id = 686) # Gemeente Eindhoven
 
-api_client = notubiz.ApiClient(configuration)
+api_client = ApiClient(configuration)
 
-speakers = notubiz.api.NotubizSpeakers(api_client).get()
+speakers = SpeakersClient(api_client).get()
 
 speaker = speakers.find_by_person_id(194366)
