--- notubiz/api/__init__.py
@@ -1,2 +1,0 @@
-from notubiz.api.speakers import NotubizSpeakers
-from notubiz.api.meeting import NotubizMeeting
