+++ notubiz/api/clients/__init__.py
@@ -1,0 +1,3 @@
+from notubiz.api.clients.events_client import EventsClient
+from notubiz.api.clients.meeting_client import MeetingClient
+from notubiz.api.clients.speakers_client import SpeakersClient
