+++ notubiz/api/dataclasses/__init__.py
@@ -1,0 +1,5 @@
+from notubiz.api.dataclasses.event import Event
+from notubiz.api.dataclasses.meeting import Meeting
+from notubiz.api.dataclasses.document import Document
+from notubiz.api.dataclasses.agenda_item import AgendaItem
+from notubiz.api.dataclasses.speakers import Speaker, Speakers
